requirements.txt
Django==4.0.1
djangorestframework==3.12.4